+++ app/keepers/keepers.go
@@ -496,9 +496,6 @@ func (ak *AppKeepers) InitKeepers(
 		runtime.NewKVStoreService(keys[btcstakingtypes.StoreKey]),
 		&btclightclientKeeper,
 		&btcCheckpointKeeper,
-		// setting the finality keeper as nil for now
-		// need to set it after finality keeper is initiated
-		nil,
 		&ak.IncentiveKeeper,
 		btcNetParams,
 		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
@@ -513,11 +510,6 @@ func (ak *AppKeepers) InitKeepers(
 		ak.CheckpointingKeeper,
 		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
 	)
-	ak.BTCStakingKeeper = *ak.BTCStakingKeeper.SetHooks(btcstakingtypes.NewMultiBtcStakingHooks(ak.FinalityKeeper.Hooks()))
-	ak.FinalityKeeper = *ak.FinalityKeeper.SetHooks(finalitytypes.NewMultiFinalityHooks(ak.BTCStakingKeeper.Hooks()))
-	// TODO this introduces circular dependency between the finality module and
-	// the btcstaking modules, need refactoring
-	ak.BTCStakingKeeper.FinalityKeeper = ak.FinalityKeeper
 
 	// create evidence keeper with router
 	evidenceKeeper := evidencekeeper.NewKeeper(
