+++ proto/babylon/finality/v1/genesis.proto
@@ -4,6 +4,7 @@ package babylon.finality.v1;
 import "gogoproto/gogo.proto";
 import "babylon/finality/v1/params.proto";
 import "babylon/finality/v1/finality.proto";
+import "babylon/finality/v1/incentive.proto";
 
 option go_package = "github.com/babylonlabs-io/babylon/x/finality/types";
 
@@ -11,22 +12,46 @@ option go_package = "github.com/babylonlabs-io/babylon/x/finality/types";
 message GenesisState {
   // params the current params of the state.
   Params params = 1 [(gogoproto.nullable) = false];
+  // voting_powers the voting power of every finality provider at every block height.
+  repeated VotingPowerFP voting_powers = 2;
+  // vp_dst_cache is the table of all providers voting power with the total at one specific block.
+  // TODO: remove this after not storing in the keeper store it anymore.
+  repeated VotingPowerDistCacheBlkHeight vp_dst_cache = 3;
   // indexed_blocks all the btc blocks and if their status are finalized.
-  repeated IndexedBlock indexed_blocks = 2;
+  repeated IndexedBlock indexed_blocks = 4;
   // evidences all the evidences ever registered.
-  repeated Evidence evidences = 3;
+  repeated Evidence evidences = 5;
   // votes_sigs contains all the votes of finality providers ever registered.
-  repeated VoteSig vote_sigs = 4;
+  repeated VoteSig vote_sigs = 6;
   // public_randomness contains all the public randomness ever committed from the finality providers.
-  repeated PublicRandomness public_randomness = 5;
+  repeated PublicRandomness public_randomness = 7;
   // pub_rand_commit contains all the public randomness commitment ever committed from the finality providers.
-  repeated PubRandCommitWithPK pub_rand_commit = 6;
+  repeated PubRandCommitWithPK pub_rand_commit = 8;
   // signing_infos represents a map between finality provider public key and their
   // signing infos.
-  repeated SigningInfo signing_infos = 7 [(gogoproto.nullable) = false];
+  repeated SigningInfo signing_infos = 9 [(gogoproto.nullable) = false];
   // missed_blocks represents a map between finality provider public key and their
   // missed blocks.
-  repeated FinalityProviderMissedBlocks missed_blocks = 8 [(gogoproto.nullable) = false];
+  repeated FinalityProviderMissedBlocks missed_blocks = 10 [(gogoproto.nullable) = false];
+}
+
+// VotingPowerFP contains the information about the voting power
+// of an finality provider in a specific block height.
+message VotingPowerFP {
+  // block_height is the height of the block the voting power was stored.
+  uint64 block_height = 1;
+  // fp_btc_pk the finality provider btc public key.
+  bytes fp_btc_pk = 2 [ (gogoproto.customtype) = "github.com/babylonlabs-io/babylon/types.BIP340PubKey" ];
+  // voting_power is the power of the finality provider at this specific block height.
+  uint64 voting_power = 3;
+}
+
+// VotingPowerDistCacheBlkHeight the total voting power of the finality providers at one specific block height
+message VotingPowerDistCacheBlkHeight {
+  // block_height is the height of the block the voting power distribution cached was stored.
+  uint64 block_height = 1;
+  // vp_distribution the finality providers distribution cache at that height.
+  VotingPowerDistCache vp_distribution = 2;
 }
 
 // VoteSig the vote of an finality provider
