+++ proto/babylon/finality/v1/params.proto
@@ -11,14 +11,16 @@ option go_package = "github.com/babylonlabs-io/babylon/x/finality/types";
 // Params defines the parameters for the module.
 message Params {
   option (gogoproto.goproto_stringer) = false;
+  // max_active_finality_providers is the maximum number of active finality providers in the BTC staking protocol
+  uint32 max_active_finality_providers = 1;
   // signed_blocks_window defines the size of the sliding window for tracking finality provider liveness
-  int64 signed_blocks_window  = 1;
+  int64 signed_blocks_window  = 2;
   // finality_sig_timeout defines how much time (in terms of blocks) finality providers have to cast a finality
   // vote before being judged as missing their voting turn on the given block
-  int64 finality_sig_timeout = 2;
+  int64 finality_sig_timeout = 3;
   // min_signed_per_window defines the minimum number of blocks that a finality provider is required to sign
   // within the sliding window to avoid being jailed
-  bytes min_signed_per_window = 3 [
+  bytes min_signed_per_window = 4 [
     (cosmos_proto.scalar)  = "cosmos.Dec",
     (gogoproto.customtype) = "cosmossdk.io/math.LegacyDec",
     (gogoproto.nullable)   = false,
@@ -26,8 +28,8 @@ message Params {
   ];
   // min_pub_rand is the minimum number of public randomness each
   // message should commit
-  uint64 min_pub_rand = 4;
+  uint64 min_pub_rand = 5;
   // jail_duration is the minimum period of time that a finality provider remains jailed
-  google.protobuf.Duration jail_duration = 5
+  google.protobuf.Duration jail_duration = 6
   [(gogoproto.nullable) = false, (amino.dont_omitempty) = true, (gogoproto.stdduration) = true];
 }
