+++ x/btcstaking/keeper/keeper.go
@@ -19,12 +19,9 @@ type (
 		cdc          codec.BinaryCodec
 		storeService corestoretypes.KVStoreService
 
-		btclcKeeper    types.BTCLightClientKeeper
-		btccKeeper     types.BtcCheckpointKeeper
-		FinalityKeeper types.FinalityKeeper
-		iKeeper        types.IncentiveKeeper
-
-		hooks types.BtcStakingHooks
+		btclcKeeper types.BTCLightClientKeeper
+		btccKeeper  types.BtcCheckpointKeeper
+		iKeeper     types.IncentiveKeeper
 
 		btcNet *chaincfg.Params
 		// the address capable of executing a MsgUpdateParams message. Typically, this
@@ -39,7 +36,6 @@ func NewKeeper(
 
 	btclcKeeper types.BTCLightClientKeeper,
 	btccKeeper types.BtcCheckpointKeeper,
-	finalityKeeper types.FinalityKeeper,
 	iKeeper types.IncentiveKeeper,
 
 	btcNet *chaincfg.Params,
@@ -49,27 +45,13 @@ func NewKeeper(
 		cdc:          cdc,
 		storeService: storeService,
 
-		btclcKeeper:    btclcKeeper,
-		btccKeeper:     btccKeeper,
-		FinalityKeeper: finalityKeeper,
-		iKeeper:        iKeeper,
-
-		hooks: nil,
+		btclcKeeper: btclcKeeper,
+		btccKeeper:  btccKeeper,
+		iKeeper:     iKeeper,
 
 		btcNet:    btcNet,
 		authority: authority,
 	}
-}
-
-// SetHooks sets the BTC staking hooks
-func (k *Keeper) SetHooks(sh types.BtcStakingHooks) *Keeper {
-	if k.hooks != nil {
-		panic("cannot set BTC staking hooks twice")
-	}
-
-	k.hooks = sh
-
-	return k
 }
 
 func (k Keeper) Logger(ctx sdk.Context) log.Logger {
