+++ x/btcstaking/keeper/query_params_test.go
@@ -10,7 +10,7 @@ import (
 )
 
 func TestParamsQuery(t *testing.T) {
-	keeper, ctx := testkeeper.BTCStakingKeeper(t, nil, nil, nil, nil)
+	keeper, ctx := testkeeper.BTCStakingKeeper(t, nil, nil, nil)
 	params := types.DefaultParams()
 
 	err := keeper.SetParams(ctx, params)
@@ -22,7 +22,7 @@ func TestParamsQuery(t *testing.T) {
 }
 
 func TestParamsByVersionQuery(t *testing.T) {
-	keeper, ctx := testkeeper.BTCStakingKeeper(t, nil, nil, nil, nil)
+	keeper, ctx := testkeeper.BTCStakingKeeper(t, nil, nil, nil)
 
 	// starting with `1` as BTCStakingKeeper creates params with version 0
 	params1 := types.DefaultParams()
