+++ x/btcstaking/types/query.go
@@ -58,7 +58,7 @@ func (ud *BTCUndelegation) ToResponse() (resp *BTCUndelegationResponse) {
 }
 
 // NewFinalityProviderResponse creates a new finality provider response based on the finaliny provider and his voting power.
-func NewFinalityProviderResponse(f *FinalityProvider, bbnBlockHeight, votingPower uint64) *FinalityProviderResponse {
+func NewFinalityProviderResponse(f *FinalityProvider, bbnBlockHeight uint64) *FinalityProviderResponse {
 	return &FinalityProviderResponse{
 		Description:          f.Description,
 		Commission:           f.Commission,
@@ -69,6 +69,5 @@ func NewFinalityProviderResponse(f *FinalityProvider, bbnBlockHeight, votingPowe
 		SlashedBtcHeight:     f.SlashedBtcHeight,
 		Jailed:               f.Jailed,
 		Height:               bbnBlockHeight,
-		VotingPower:          votingPower,
 	}
 }
