+++ x/finality/client/cli/query.go
@@ -29,6 +29,9 @@ func GetQueryCmd(queryRoute string) *cobra.Command {
 	}
 
 	cmd.AddCommand(CmdQueryParams())
+	cmd.AddCommand(CmdActiveFinalityProvidersAtHeight())
+	cmd.AddCommand(CmdFinalityProviderPowerAtHeight())
+	cmd.AddCommand(CmdActivatedHeight())
 	cmd.AddCommand(CmdListPublicRandomness())
 	cmd.AddCommand(CmdListPubRandCommit())
 	cmd.AddCommand(CmdBlock())
@@ -37,6 +40,99 @@ func GetQueryCmd(queryRoute string) *cobra.Command {
 	cmd.AddCommand(CmdListEvidences())
 	cmd.AddCommand(CmdSigningInfo())
 	cmd.AddCommand(CmdAllSigningInfo())
+
+	return cmd
+}
+
+func CmdFinalityProviderPowerAtHeight() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "finality-provider-power-at-height [fp_btc_pk_hex] [height]",
+		Short: "get the voting power of a given finality provider at a given height",
+		Args:  cobra.ExactArgs(2),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx := client.GetClientContextFromCmd(cmd)
+
+			queryClient := types.NewQueryClient(clientCtx)
+
+			height, err := strconv.ParseUint(args[1], 10, 64)
+			if err != nil {
+				return err
+			}
+			res, err := queryClient.FinalityProviderPowerAtHeight(cmd.Context(), &types.QueryFinalityProviderPowerAtHeightRequest{
+				FpBtcPkHex: args[0],
+				Height:     height,
+			})
+			if err != nil {
+				return err
+			}
+
+			return clientCtx.PrintProto(res)
+		},
+	}
+
+	flags.AddQueryFlagsToCmd(cmd)
+
+	return cmd
+}
+
+func CmdActivatedHeight() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "activated-height",
+		Short: "get activated height, i.e., the first height where there exists 1 finality provider with voting power",
+		Args:  cobra.NoArgs,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx := client.GetClientContextFromCmd(cmd)
+
+			queryClient := types.NewQueryClient(clientCtx)
+
+			res, err := queryClient.ActivatedHeight(cmd.Context(), &types.QueryActivatedHeightRequest{})
+			if err != nil {
+				return err
+			}
+
+			return clientCtx.PrintProto(res)
+		},
+	}
+
+	flags.AddQueryFlagsToCmd(cmd)
+
+	return cmd
+}
+
+func CmdActiveFinalityProvidersAtHeight() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "active-finality-providers-at-height [height]",
+		Short: "retrieve all active finality providers at a given babylon height",
+		Args:  cobra.ExactArgs(1),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx := client.GetClientContextFromCmd(cmd)
+
+			queryClient := types.NewQueryClient(clientCtx)
+
+			height, err := strconv.ParseUint(args[0], 10, 64)
+			if err != nil {
+				return err
+			}
+
+			pageReq, err := client.ReadPageRequest(cmd.Flags())
+			if err != nil {
+				return err
+			}
+
+			res, err := queryClient.ActiveFinalityProvidersAtHeight(cmd.Context(), &types.QueryActiveFinalityProvidersAtHeightRequest{
+				Height:     height,
+				Pagination: pageReq,
+			})
+			if err != nil {
+				return err
+			}
+
+			return clientCtx.PrintProto(res)
+		},
+	}
+
+	flags.AddQueryFlagsToCmd(cmd)
+	flags.AddPaginationFlagsToCmd(cmd, "finality-providers-at-height")
 
 	return cmd
 }
