+++ x/finality/types/expected_keepers.go
@@ -3,7 +3,6 @@ package types
 import (
 	"context"
 
-	bbn "github.com/babylonlabs-io/babylon/types"
 	bstypes "github.com/babylonlabs-io/babylon/x/btcstaking/types"
 	etypes "github.com/babylonlabs-io/babylon/x/epoching/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -19,6 +18,7 @@ type BTCStakingKeeper interface {
 	GetBTCStakingActivatedHeight(ctx context.Context) (uint64, error)
 	GetVotingPowerDistCache(ctx context.Context, height uint64) (*bstypes.VotingPowerDistCache, error)
 	RemoveVotingPowerDistCache(ctx context.Context, height uint64)
+	JailFinalityProvider(ctx context.Context, fpBTCPK []byte) error
 	UnjailFinalityProvider(ctx context.Context, fpBTCPK []byte) error
 }
 
@@ -32,12 +32,4 @@ type CheckpointingKeeper interface {
 type IncentiveKeeper interface {
 	RewardBTCStaking(ctx context.Context, height uint64, filteredDc *bstypes.VotingPowerDistCache)
 	IndexRefundableMsg(ctx context.Context, msg sdk.Msg)
-}
-
-type BtcStakingHooks interface {
-	AfterFinalityProviderActivated(ctx context.Context, btcPk *bbn.BIP340PubKey) error
-}
-
-type FinalityHooks interface {
-	AfterSluggishFinalityProviderDetected(ctx context.Context, btcPk *bbn.BIP340PubKey) error
 }
